feat(widget): moved some part of logic to init widget hook
