--- src/components/screens/widget/useSelectToken.ts
@@ -1,1 +1,0 @@
-export const useSelectToken = () => {};
